sessions, authentication and cookies

+++ delete-article.php
@@ -8,14 +8,11 @@ $conn = getDB();
 
 if (isset($_GET['id'])) {
 
-  $article = getArticle($conn, $_GET['id']);
+  $article = getArticle($conn, $_GET['id'], 'id');
 
   if ($article) {
 
     $id = $article['id'];
-    $title = $article['title'];
-    $content = $article['content'];
-    $published_at = $article['published_at'];
 
   } else {
       die("article not found");
@@ -25,7 +22,7 @@ if (isset($_GET['id'])) {
 } else {
   die("id not supplied, article not found");
 }
-
+if ($_SERVER["REQUEST_METHOD"] == "POST") {
   $sql = "DELETE FROM articles
           WHERE id = ?";
 
@@ -33,16 +30,29 @@ $stmt = mysqli_prepare($conn, $sql);
 
 if ($stmt === false) {
   echo mysqli_error($conn);
-} else {
+  } else {
 
   mysqli_stmt_bind_param($stmt, "i", $id );
 
-if (mysqli_stmt_execute($stmt)) {
+  if (mysqli_stmt_execute($stmt)) {
 
 
   redirect("/phptutorial/index.php");
-} else {
-  echo mysqli_stmt_error($stmt);
-}
+  } else {
+    echo mysqli_stmt_error($stmt);
+   }
 
-}  die("form is valid");
+  }  
+}
+?>
+
+<?php require 'includes/header.php'; ?>
+
+<h2>Delete Article</h2>
+
+<form method="post">
+        <button>Delete</button>        
+        <a href="article.php?id=<?= $article['id']; ?>">Cancel</a>
+      </form>
+
+<?php require 'includes/footer.php'; ?>
